+++ routes/index.js
@@ -1,9 +1,12 @@
 const express = require('express');
 const router = express.Router();
+const globals = require('../lib/globals');
+
 
 /* GET home page. */
 router.get('/', function(req, res) {
-  res.render('index', { title: 'Trivernis.net' });
+  let info = req.session.acceptedCookies? null: globals.cookieInfo;
+  res.render('index', { title: 'Trivernis.net', info: info});
 });
 
 module.exports = router;